requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
django-ratelimit==4.1.0

# Production Requirements
gunicorn==21.2.0
uvicorn==0.27.1
whitenoise==6.6.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0